Render images in random locations
I set up a useEffect hook that shuffles the images and updates the state
of the flags. The flags are then sent as props to the Card components to
be set as the src attribute, which then displays the flags in a random
order in the UI when the app starts. Here are two issues that prevented
me from completing this part of the project in a timely manner:

1. I originally had the images placed in the src folder of this project.
This was causing issues with some of the image urls, which meant that
they wouldn't load. I therefor moved the images to the public folder,
and made use of relative path urls in my images array, rather than
importing the images in app.js and making references to those urls with
keywords.

2. An annoying bug I had to fix was that the images would load when the
app started, but would not load when the app was refreshed. The reason
for this was because in my useEffect hook (in my shuffleFlags function),
I made the new constant variable newFlags equal to the reference of the
flagsDefaultState object:

const newFlags = flagsDefaultState;

rather than making it equal to a shallow copy of the flagsDefaultState:

const newFlags = { ...flagsDefaultState };

This meant that any changes made to the newFLags object were also made
to the flagsDefaultState object, which is undesired in this program.
This is a lesson that I need to remember: make shallow copies of
objects! Don't point them to references of existing objects. A simple
mistake, but deadly.