Rework state logic to add points
I reworked the state logic so that, other than the flags state variable,
there is only one other state variable; the gameControl state object.
This object contains all of the necessary game statistics;
selectedFlags, and score. It also now contains the shuffleTrigger
boolean variable, which is now the only dependency variable that is
being tracked by the useEffect hook. The handleClick function is what
handles the logic behind adding to the score, or resetting the score to
0, and the useEffect hook is still responsible for calling the
shuffleFlags function and re-rendering the flags.