Add some final styling features
Here is a list of some final styling features that I added to complete
this project:

1. Add borders to images

2. Add box-shadow on image hover

3. Add grow from center image animation, and associate that animation
with a class called animation, that is conditionally set to all card
components based on a new state variable “imagesRendered”

4. Add loading screen. This loading screen is controlled from the
useEffect hook, in tandem with the new loadPage state variable. loadPage
is set to true when the page first loads. Then, it is set to false after
a certain time delay in the useEffect hook. It is not set to true after
that.